refactor: deduplicate printing logic in WebContents
